Additional Event Plumbings and contact detection
- Additional rawEvent plumbing so that each test (ours as well)
receives raw event data in order to be able to query for lower, more
complicated case.
- Added a new method findBody() in order detect the contact. That is,
finding out the body that we clicked on for example. This seems to be
better method, since this implementation is relying on ContactManager
that’s provided by Box2d.
- Experiment of new gesture of ‘Ctrl + Click’ on a vertex so we know
user wants to begin creating a vertex. To be continued on this one.