Optimize pair buffer in DefaultBroadPhaseBuffer (#64)
Optimize pair buffer in DefaultBroadPhaseBuffer to be an array of  'long' instead of objects.

This changes pair buffer to encode the pair in a long instead of an object on the heap. This improves sorting speed, which is quite a bottleneck if there are many pairs. Additionally, it avoids a few object allocations, but that's not the main point here.

Pair is deleted, because it is no longer used.